backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
web3>=6.19.0
openai>=1.30.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
pydantic>=2.9.0
httpx>=0.27.0
